requirements.txt
# Core dependencies
ollama==0.4.4
openai==1.12.0
python-dotenv==1.0.1
numpy==1.26.4
requests==2.31.0

# Messaging (Discord/WhatsApp)
flask==3.0.0
discord.py-self==2.0.0
asyncio

# Audio and TTS
sounddevice==0.4.7
kokoro==0.9.4

# GUI automation and image processing
pyautogui==0.9.54
Pillow==10.3.0

# Spotify integration
spotipy==2.23.0

# Context tools - clipboard, window detection, web scraping
pyperclip==1.8.2
pygetwindow==0.0.9
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# For PyTorch with CUDA 12.1 support, install separately with:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# Compatible dependency versions to avoid conflicts
pydantic==2.9.2
pydantic-core==2.23.4
typing-extensions==4.12.2
httpx==0.27.2
